Internal console: Move show_completion_list method to it
That method is no longer used in TextEditBaseWidget
